Move exports from animationParser
