+++ src/reanimated2/layoutReanimation/web/animationsManager.ts
@@ -1,11 +1,15 @@
 'use strict';
 
-import type { AnimationConfig, AnimationNames, CustomConfig } from './config';
+import type {
+  AnimationConfig,
+  AnimationNames,
+  CustomConfig,
+  TransitionData,
+} from './config';
 import { Animations } from './config';
 import type { AnimatedComponentProps } from '../../../createAnimatedComponent/utils';
 import { LayoutAnimationType } from '../animationBuilder/commonTypes';
-import type { TransitionData } from './animationParser';
-import type { StyleProps } from '../..//commonTypes';
+import type { StyleProps } from '../../commonTypes';
 import { createAnimationWithExistingTransform } from './createAnimation';
 import {
   getProcessedConfig,
