+++ src/reanimated2/layoutReanimation/web/transition/Fading.web.ts
@@ -1,5 +1,5 @@
 'use strict';
-import type { TransitionData } from '../animationParser';
+import type { TransitionData } from '../config';
 
 export function FadingTransition(name: string, transitionData: TransitionData) {
   const { translateX, translateY, scaleX, scaleY } = transitionData;
