+++ src/reanimated2/layoutReanimation/web/transition/Linear.web.ts
@@ -1,5 +1,5 @@
 'use strict';
-import type { TransitionData } from '../animationParser';
+import type { TransitionData } from '../config';
 
 export function LinearTransition(name: string, transitionData: TransitionData) {
   const { translateX, translateY, scaleX, scaleY } = transitionData;
